upgrade to last SvelteKit version before routes change
