+++ web/src/app.html
@@ -5,9 +5,11 @@
     <link rel="icon" href="/favicon.png" />
     <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    %svelte.head%
+    %sveltekit.head%
   </head>
   <body>
-    %svelte.body%
+    <div id="ergo">
+      %sveltekit.body%
+    </div>
   </body>
 </html>
