+++ web/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "./.svelte-kit/tsconfig.json",
   "compilerOptions": {
     "moduleResolution": "node",
     "module": "esnext",
@@ -13,10 +14,7 @@
     "baseUrl": ".",
     "allowJs": true,
     "checkJs": true,
-    "strict": true,
-    "paths": {
-      "$lib/*": ["src/lib/*"]
-    }
+    "strict": true
   },
   "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.ts", "src/**/*.svelte"]
 }
